app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
redis==5.0.1
pypdf==3.17.1
google-generativeai==0.3.1
python-dotenv==1.0.0
pdf2image==1.16.3
Pillow==10.1.0 